% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_new_method_stability.R,
%   R/reproduce_stability.R
\name{calculate_jaccard_internal}
\alias{calculate_jaccard_internal}
\title{Calculate Jaccard Index - Internal}
\usage{
calculate_jaccard_internal(network1, network2)

calculate_jaccard_internal(network1, network2)
}
\description{
Calculate Jaccard Index - Internal

Calculate Jaccard Index - Internal
}
\keyword{internal}
